backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.17.0
opencv-python==4.8.1.78
numpy==1.26.4
Pillow==10.1.0
PyJWT==2.8.0
werkzeug==3.0.1
scikit-learn==1.3.2
